fix: return legacy did:web document when resolving legacy did Signed-off-by: Ariel Gentile <gentilester@gmail.com>
