+++ apps/vs-agent/src/utils/CachedWebDidResolver.ts
@@ -7,7 +7,16 @@ import {
 } from '@credo-ts/core'
 import { ParsedDID } from 'did-resolver'
 
+import { getLegacyDidDocument } from './legacyDidWeb'
+
 export class CachedWebDidResolver extends WebDidResolver {
+  private publicApiBaseUrl: string
+
+  public constructor(options: { publicApiBaseUrl: string }) {
+    super()
+    this.publicApiBaseUrl = options.publicApiBaseUrl
+  }
+
   public async resolve(
     agentContext: AgentContext,
     did: string,
@@ -17,10 +26,8 @@ export class CachedWebDidResolver extends WebDidResolver {
     // First check within our own public dids, as there is no need to resolve it through HTTPS
     const didRepository = agentContext.dependencyManager.resolve(DidRepository)
     const didRecord = await didRepository.findSingleByQuery(agentContext, {
-      $or: [
-        { did, method: 'web' },
-        { domain: parsed.id, method: 'webvh' }, // Find equivalent did:webvh, since this might be a legacy alias
-      ],
+      did,
+      method: 'web',
     })
 
     if (didRecord?.didDocument) {
@@ -28,6 +35,23 @@ export class CachedWebDidResolver extends WebDidResolver {
         didDocument: didRecord.didDocument,
         didDocumentMetadata: {},
         didResolutionMetadata: {},
+      }
+    }
+
+    // Find equivalent did:webvh, since this might be a legacy alias
+    const webVhdDidRecord = await didRepository.findSingleByQuery(agentContext, {
+      domain: parsed.id,
+      method: 'webvh',
+    })
+
+    if (webVhdDidRecord?.didDocument) {
+      const legacyDidDocument = getLegacyDidDocument(webVhdDidRecord.didDocument, this.publicApiBaseUrl)
+      if (legacyDidDocument) {
+        return {
+          didDocument: legacyDidDocument,
+          didDocumentMetadata: {},
+          didResolutionMetadata: {},
+        }
       }
     }
